third_party/blink/web_tests/platform/generic/http/tests/websocket/permessage-deflate-comp-bit-onoff-expected.txt

Test compression enabled/disabled frame receiving.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending message: "Hello"
PASS event.data is 'Hello'
Sending message: "DisableCompression"
PASS event.data is 'DisableCompression'
Sending message: "World"
PASS event.data is 'World'
Sending message: "EnableCompression"
PASS event.data is 'EnableCompression'
Sending message: ""
PASS event.data is ''
Sending message: ""
PASS event.data is ''
Sending message: "Goodbye"
PASS event.data is 'Goodbye'
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
